refactor: docx converter
Refactored DOCX converter internals for better maintainability
