+++ contextgem/internal/converters/__init__.py
@@ -1,0 +1,37 @@
+#
+# ContextGem
+#
+# Copyright 2025 Shcherbak AI AS. All rights reserved. Developed by Sergii Shcherbak.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+from contextgem.internal.converters.docx import (
+    WORD_XML_NAMESPACES,
+    DocxContentError,
+    DocxConverterError,
+    DocxFormatError,
+    DocxXmlError,
+    _DocxConverterBase,
+    _DocxPackage,
+)
+
+__all__ = [
+    "WORD_XML_NAMESPACES",
+    "DocxConverterError",
+    "DocxFormatError",
+    "DocxXmlError",
+    "DocxContentError",
+    "_DocxConverterBase",
+    "_DocxPackage",
+]
