+++ contextgem/internal/converters/docx/__init__.py
@@ -1,0 +1,37 @@
+#
+# ContextGem
+#
+# Copyright 2025 Shcherbak AI AS. All rights reserved. Developed by Sergii Shcherbak.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+from contextgem.internal.converters.docx.base import _DocxConverterBase
+from contextgem.internal.converters.docx.exceptions import (
+    DocxContentError,
+    DocxConverterError,
+    DocxFormatError,
+    DocxXmlError,
+)
+from contextgem.internal.converters.docx.namespaces import WORD_XML_NAMESPACES
+from contextgem.internal.converters.docx.package import _DocxPackage
+
+__all__ = [
+    "_DocxConverterBase",
+    "DocxConverterError",
+    "DocxFormatError",
+    "DocxXmlError",
+    "DocxContentError",
+    "WORD_XML_NAMESPACES",
+    "_DocxPackage",
+]
