+++ contextgem/internal/converters/docx/namespaces.py
@@ -1,0 +1,38 @@
+#
+# ContextGem
+#
+# Copyright 2025 Shcherbak AI AS. All rights reserved. Developed by Sergii Shcherbak.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+"""
+Namespaces for the DOCX converter module.
+
+This module defines the XML namespaces used in DOCX files for different elements.
+It provides a dictionary of namespace URIs mapped to their prefixes, which are
+used to parse and process the XML content of DOCX files.
+"""
+
+
+# Define XML namespaces used in DOCX files
+WORD_XML_NAMESPACES = {
+    "w": "http://schemas.openxmlformats.org/wordprocessingml/2006/main",
+    "wp": "http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing",
+    "a": "http://schemas.openxmlformats.org/drawingml/2006/main",
+    "pic": "http://schemas.openxmlformats.org/drawingml/2006/picture",
+    "r": "http://schemas.openxmlformats.org/officeDocument/2006/relationships",
+    "rels": "http://schemas.openxmlformats.org/package/2006/relationships",
+    "v": "urn:schemas-microsoft-com:vml",
+    "mc": "http://schemas.openxmlformats.org/markup-compatibility/2006",
+}
