+++ docs/docs-raw-for-llm.txt
@@ -5651,7 +5651,7 @@ Converters
 
 class contextgem.public.converters.DocxConverter
 
-   Bases: "object"
+   Bases: "_DocxConverterBase"
 
    Converter for DOCX files into ContextGem documents.
